- DONE: 	- Refactored whole code dividing into modules - TODO: 	- Add login/signup future 	- Remove `review2 or change the constructor value
