+++ stay_safe/lib/const.dart
@@ -1,0 +1,7 @@
+//! This file is used to show Reviews for the given place
+//! TODO: Make review2 and review file into one
+//!
+//! authors @rooknpown
+
+String lat = "36.37379078760264";
+String lng = "127.35905994710093";
