+++ stay_safe/lib/db/database.dart
@@ -1,3 +1,8 @@
+//! This file is used to show Reviews for the given place
+//! TODO: Make review2 and review file into one
+//!
+//! authors @castlecowrice
+
 import 'dart:convert';
 import "package:mongo_dart/mongo_dart.dart";
