+++ stay_safe/lib/login_page.dart
@@ -1,0 +1,4 @@
+//! This file is used to show Reviews for the given place
+//! TODO: Make review2 and review file into one
+//!
+//! authors @kalo
