+++ stay_safe/lib/main.dart
@@ -1,17 +1,21 @@
+//! This file is used to show Reviews for the given place
+//! TODO: Make review2 and review file into one
+//!
+//! authors @rooknpown
+
 import 'package:flutter/material.dart';
 import 'package:google_maps_place_picker/google_maps_place_picker.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:http/http.dart' as http;
+import './homepage.dart';
 import 'dart:convert';
+import './const.dart';
 
 // Your api key storage.
 // import 'keys.dart';
 
 void main() => runApp(MyApp());
-
-String lat = "36.37379078760264";
-String lng = "127.35905994710093";
 
 class MyApp extends StatelessWidget {
   // Light Theme
@@ -47,1035 +51,6 @@ class MyApp extends StatelessWidget {
       home: HomePage(),
       debugShowCheckedModeBanner: false,
     );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  const HomePage({Key key}) : super(key: key);
-
-  static final kInitialPosition = LatLng(36.37379078760264, 127.35905994710093);
-
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  PickResult selectedPlace;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text("Stay Safe"),
-        ),
-        bottomNavigationBar: BottomAppBar(
-          child: new Row(
-            mainAxisSize: MainAxisSize.max,
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              IconButton(
-                icon: Icon(Icons.home),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => HomeScreen()),
-                  );
-                },
-              ),
-              IconButton(
-                icon: Icon(Icons.search),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) {
-                        return Scaffold(
-                            appBar: AppBar(
-                              title: Text('Search Places. Find Safety'),
-                              automaticallyImplyLeading: false,
-                            ),
-                            body: PlacePicker(
-                              apiKey: "AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc",
-                              initialPosition: HomePage.kInitialPosition,
-                              useCurrentLocation: true,
-                              selectInitialPosition: true,
-
-                              //usePlaceDetailSearch: true,
-                              onPlacePicked: (result) {
-                                selectedPlace = result;
-                                Navigator.of(context).pop();
-                                setState(() {});
-                              },
-                              //forceSearchOnZoomChanged: true,
-                              //automaticallyImplyAppBarLeading: false,
-                              //autocompleteLanguage: "ko",
-                              //region: 'au',
-                              //selectInitialPosition: true,
-                              selectedPlaceWidgetBuilder: (_, selectedPlace,
-                                  state, isSearchBarFocused) {
-                                print(
-                                    "state: $state, isSearchBarFocused: $isSearchBarFocused" +
-                                        selectedPlace.name);
-                                return isSearchBarFocused
-                                    ? Container()
-                                    : FloatingCard(
-                                        bottomPosition:
-                                            0.0, // MediaQuery.of(context) will cause rebuild. See MediaQuery document for the information.
-                                        leftPosition: 0.0,
-                                        rightPosition: 0.0,
-                                        width: 500,
-                                        borderRadius:
-                                            BorderRadius.circular(12.0),
-                                        child: state == SearchingState.Searching
-                                            ? Center(
-                                                child:
-                                                    CircularProgressIndicator())
-                                            : RaisedButton(
-                                                child: Text("Pick Here"),
-                                                onPressed: () {
-                                                  // IMPORTANT: You MUST manage selectedPlace data yourself as using this build will not invoke onPlacePicker as
-                                                  //            this will override default 'Select here' Button.
-                                                  print(
-                                                      "do something with [selectedPlace] data");
-                                                  print(selectedPlace.placeId);
-                                                  print(selectedPlace.rating);
-                                                  Navigator.push(
-                                                    context,
-                                                    MaterialPageRoute(
-                                                      builder: (context) =>
-                                                          SearchResult(
-                                                              selectedPlace:
-                                                                  selectedPlace),
-                                                    ),
-                                                  );
-                                                  //Navigator.of(context).pop();
-                                                },
-                                              ),
-                                      );
-                              },
-                              // pinBuilder: (context, state) {
-                              //   if (state == PinState.Idle) {
-                              //     return Icon(Icons.favorite_border);
-                              //   } else {
-                              //     return Icon(Icons.favorite);
-                              //   }
-                              // },
-                            ));
-                      },
-                    ),
-                  );
-                },
-              ),
-              IconButton(
-                icon: Icon(Icons.add),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) {
-                        return Scaffold(
-                            appBar: AppBar(
-                              title: Text('Search Places. Add Review'),
-                              automaticallyImplyLeading: false,
-                            ),
-                            body: PlacePicker(
-                              apiKey: "AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc",
-                              initialPosition: HomePage.kInitialPosition,
-                              useCurrentLocation: true,
-                              selectInitialPosition: true,
-
-                              //usePlaceDetailSearch: true,
-                              onPlacePicked: (result) {
-                                selectedPlace = result;
-                                Navigator.of(context).pop();
-                                setState(() {});
-                              },
-                              //forceSearchOnZoomChanged: true,
-                              //automaticallyImplyAppBarLeading: false,
-                              //autocompleteLanguage: "ko",
-                              //region: 'au',
-                              //selectInitialPosition: true,
-                              selectedPlaceWidgetBuilder: (_, selectedPlace,
-                                  state, isSearchBarFocused) {
-                                print(
-                                    "state: $state, isSearchBarFocused: $isSearchBarFocused");
-                                return isSearchBarFocused
-                                    ? Container()
-                                    : FloatingCard(
-                                        bottomPosition:
-                                            0.0, // MediaQuery.of(context) will cause rebuild. See MediaQuery document for the information.
-                                        leftPosition: 0.0,
-                                        rightPosition: 0.0,
-                                        width: 500,
-                                        borderRadius:
-                                            BorderRadius.circular(12.0),
-                                        child: state == SearchingState.Searching
-                                            ? Center(
-                                                child:
-                                                    CircularProgressIndicator())
-                                            : RaisedButton(
-                                                child: Text("Pick Here"),
-                                                onPressed: () {
-                                                  // IMPORTANT: You MUST manage selectedPlace data yourself as using this build will not invoke onPlacePicker as
-                                                  //            this will override default 'Select here' Button.
-                                                  print(
-                                                      "do something with [selectedPlace] data");
-                                                  print(selectedPlace.placeId);
-                                                  print(selectedPlace.rating);
-                                                  Navigator.push(
-                                                    context,
-                                                    MaterialPageRoute(
-                                                      builder: (context) =>
-                                                          MakeReview(
-                                                              selectedPlace:
-                                                                  selectedPlace),
-                                                    ),
-                                                  );
-                                                  //Navigator.of(context).pop();
-                                                },
-                                              ),
-                                      );
-                              },
-                              // pinBuilder: (context, state) {
-                              //   if (state == PinState.Idle) {
-                              //     return Icon(Icons.favorite_border);
-                              //   } else {
-                              //     return Icon(Icons.favorite);
-                              //   }
-                              // },
-                            ));
-                      },
-                    ),
-                  );
-                },
-              ),
-              // IconButton(
-              //   icon: Icon(Icons.account_box),
-              //   onPressed: () {},
-              // ),
-            ],
-          ),
-        ),
-        body: new Image.network(
-          "https://s3-alpha-sig.figma.com/img/1706/4c0e/56adf1329a844173a7b98575ec9d39ac?Expires=1607299200&Signature=GXhxZ-kWs6ClBSTzAZSiS2Pvkiuh3wnR52hyr9hGDRcO7VDN6MM5DIeZi68WrWfGLm7TJ2VTxzPbS0a2XMBJ39kyP1mZ4d-Z7TYk90f2VwBh1NLM-8BUzx8JUwsw0rTzIVS2egnK4N57oyIpXzMwzQ~-XhM-YjFv4A1ab8WlMtLU0JTXL60I~g0VDVBBQThgMGjyFJOLMqvoVHpAlrrv7t5BYvBoWIGTrZazzzZd24YNe0C4AfKEIebzs5fpzZOxN53y9hR616mcuZoFGIv4vlW0bZM2L9u27uVPm-Lnw0Dq4AwHiHZNkWgUyNp1w8VZev95gg3Cs5xsgsdh9rXOFQ__&Key-Pair-Id=APKAINTVSUGEWH5XD5UA",
-          fit: BoxFit.cover,
-          height: double.infinity,
-          width: double.infinity,
-          alignment: Alignment.center,
-        ));
-  }
-}
-
-class HomeScreen extends StatelessWidget {
-  String buildPhotoURL(String photoReference) {
-    return "https://maps.googleapis.com/maps/api/place/photo?maxwidth=1000&photoreference=${photoReference}&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    Future<List<dynamic>> info = fetchAlbum();
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Discover places"),
-      ),
-      body: FutureBuilder<List<dynamic>>(
-        future: info,
-        builder: (BuildContext context, AsyncSnapshot<List<dynamic>> snapshot) {
-          if (snapshot.hasData) {
-            List<Widget> widgetlist = new List<Widget>();
-            print("snaplength:" + snapshot.data.length.toString());
-            for (int i = 0; i < snapshot.data.length; i = i + 2) {
-              widgetlist.add(new Text(snapshot.data[i + 0],
-                  style: TextStyle(height: 1, fontSize: 25)));
-              widgetlist.add(new Expanded(
-                  child: Container(
-                      child: ConstrainedBox(
-                          constraints: BoxConstraints(maxHeight: 500.0),
-                          child: FlatButton(
-                              onPressed: () {
-                                Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                        builder: (context) => AllReviews2(
-                                                selectedPlace: [
-                                                  snapshot.data[i + 0],
-                                                  snapshot.data[i + 1]
-                                                ])));
-                              },
-                              padding: EdgeInsets.all(0.0),
-                              child: Image.network(
-                                  buildPhotoURL(snapshot.data[i + 1])))))));
-            }
-            return new ListView(
-              children: widgetlist,
-            );
-          } else if (snapshot.hasError) {
-            return ListView(
-              children: [
-                Text("Searching for places",
-                    style: TextStyle(height: 1, fontSize: 25)),
-              ],
-            );
-          } else {
-            return CircularProgressIndicator();
-          }
-        },
-      ),
-    );
-  }
-}
-
-String searchNearBy() {
-  print(lat + lng);
-  return "https://maps.googleapis.com/maps/api/place/nearbysearch/json?location=$lat,$lng&radius=100&type=restaurant&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-}
-
-Future<List<dynamic>> fetchAlbum() async {
-  final response = await http.get(searchNearBy());
-  Map<String, dynamic> user = jsonDecode(response.body);
-  var list = new List();
-  for (int i = 0; i < user["results"].length; i++) {
-    if (user["results"][i]["name"] != null &&
-        user["results"][i]["photos"] != null) {
-      list.add(user["results"][i]["name"]);
-      list.add(user["results"][i]["photos"][0]["photo_reference"]);
-    }
-  }
-  await Future.delayed(const Duration(seconds: 1), () {});
-  print(user["results"][0]["name"]);
-  return list;
-}
-
-class SearchResult extends StatelessWidget {
-  final PickResult selectedPlace;
-
-  String buildPhotoURL(String photoReference) {
-    return "https://maps.googleapis.com/maps/api/place/photo?maxwidth=1000&photoreference=${photoReference}&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-  }
-
-  SearchResult({@required this.selectedPlace});
-  @override
-  Widget build(BuildContext context) {
-    lat = selectedPlace.geometry.location.lat.toString();
-    lng = selectedPlace.geometry.location.lng.toString();
-    return Scaffold(
-        appBar: AppBar(
-          title: Text(selectedPlace.name),
-        ),
-        body: ListView(
-          children: [
-            Padding(
-                padding: EdgeInsets.only(right: 1.0),
-                child: SizedBox(
-                  height: 400,
-                  child: Image.network(
-                      buildPhotoURL(selectedPlace.photos[0].photoReference),
-                      height: 400,
-                      fit: BoxFit.fill),
-                )),
-            Text('Safety Level', style: TextStyle(height: 1, fontSize: 25)),
-            Center(
-                child: RatingBar.builder(
-              initialRating: 3,
-              minRating: 1,
-              direction: Axis.horizontal,
-              allowHalfRating: true,
-              itemCount: 5,
-              itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-              itemSize: 65.0,
-              itemBuilder: (context, _) => Icon(
-                Icons.star,
-                color: Colors.amber,
-              ),
-              onRatingUpdate: (rating) {
-                print(rating);
-              },
-            )),
-            Text('Overall Experience',
-                style: TextStyle(height: 1, fontSize: 25)),
-            Center(
-                child: RatingBar.builder(
-              initialRating: selectedPlace.rating,
-              minRating: 1,
-              direction: Axis.horizontal,
-              allowHalfRating: true,
-              itemCount: 5,
-              itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-              itemSize: 65.0,
-              itemBuilder: (context, _) => Icon(
-                Icons.star,
-                color: Colors.lightBlue,
-              ),
-              onRatingUpdate: (rating) {
-                print(rating);
-              },
-            )),
-            Center(
-              child: ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) =>
-                              AllReviews(selectedPlace: selectedPlace)));
-                },
-                child: Text('SEE ALL REVIEWS'),
-              ),
-            )
-          ],
-        ));
-  }
-}
-
-class AllReviews extends StatelessWidget {
-  final PickResult selectedPlace;
-
-  String buildPhotoURL(String photoReference) {
-    return "https://maps.googleapis.com/maps/api/place/photo?maxwidth=1000&photoreference=${photoReference}&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-  }
-
-  AllReviews({@required this.selectedPlace});
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text("Reviews: " + selectedPlace.name),
-        ),
-        body: ListView(
-          children: [
-            Padding(
-                padding: EdgeInsets.only(right: 1.0),
-                child: SizedBox(
-                  height: 200,
-                  child: Image.network(
-                      buildPhotoURL(selectedPlace.photos[0].photoReference),
-                      height: 200,
-                      fit: BoxFit.fill),
-                )),
-            Padding(
-              padding: EdgeInsets.only(top: 16.0, left: 20.0, right: 20.0),
-              child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text('Safety Level',
-                        style: TextStyle(height: 1, fontSize: 20)),
-                    Text('Overall Experience',
-                        style: TextStyle(height: 1, fontSize: 20)),
-                  ]),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: selectedPlace.rating,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          'Anonymous                                          ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text("People can't help taking off their masks while eating.",
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 2,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: selectedPlace.rating,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          'Will kenny                                           ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text(
-                      'Safe, since people are not allowed to sit face to face.',
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: selectedPlace.rating,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          '    Harry Kim                                           ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text('This place is very clean and neat!',
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 4,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: selectedPlace.rating,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-          ],
-        ));
-  }
-}
-
-class AllReviews2 extends StatelessWidget {
-  final List<String> selectedPlace;
-
-  String buildPhotoURL(String photoReference) {
-    return "https://maps.googleapis.com/maps/api/place/photo?maxwidth=1000&photoreference=${photoReference}&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-  }
-
-  AllReviews2({@required this.selectedPlace});
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text("Reviews: " + selectedPlace[0]),
-        ),
-        body: ListView(
-          children: [
-            Padding(
-                padding: EdgeInsets.only(right: 1.0),
-                child: SizedBox(
-                  height: 200,
-                  child: Image.network(buildPhotoURL(selectedPlace[1]),
-                      height: 200, fit: BoxFit.fill),
-                )),
-            Padding(
-              padding: EdgeInsets.only(top: 16.0, left: 20.0, right: 20.0),
-              child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text('Safety Level',
-                        style: TextStyle(height: 1, fontSize: 20)),
-                    Text('Overall Experience',
-                        style: TextStyle(height: 1, fontSize: 20)),
-                  ]),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          'Anonymous                                          ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text("People can't help taking off their masks while eating.",
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 2,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          'Will kenny                                           ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text(
-                      'Safe, since people are not allowed to sit face to face.',
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.account_box),
-                Column(children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Container(
-                      child: Text(
-                          '    Harry Kim                                           ',
-                          style: TextStyle(height: 3, fontSize: 20)),
-                    ),
-                  ),
-                  Text('This place is very clean and neat!',
-                      style: TextStyle(height: 1, fontSize: 15))
-                ])
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 4,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-                Center(
-                    child: RatingBar.builder(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemSize: 25.0,
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.lightBlue,
-                  ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )),
-              ],
-            ),
-          ],
-        ));
-  }
-}
-
-class MakeReview extends StatelessWidget {
-  final PickResult selectedPlace;
-  final myController = TextEditingController();
-  double safety;
-  double overall;
-  String content;
-
-  String buildPhotoURL(String photoReference) {
-    return "https://maps.googleapis.com/maps/api/place/photo?maxwidth=1000&photoreference=${photoReference}&key=AIzaSyDqOOHRnNiYaCweRNtiXVQswGAb1Pz88Yc";
-  }
-
-  MakeReview({@required this.selectedPlace});
-
-  @override
-  Widget build(BuildContext context) {
-    lat = selectedPlace.geometry.location.lat.toString();
-    lng = selectedPlace.geometry.location.lng.toString();
-    return Scaffold(
-        appBar: AppBar(
-          title: Text("Make Review on: ${selectedPlace.name}"),
-        ),
-        body: ListView(
-          children: [
-            Padding(
-                padding: EdgeInsets.only(),
-                child: SizedBox(
-                  height: 200,
-                  child: Image.network(
-                      buildPhotoURL(selectedPlace.photos[0].photoReference),
-                      height: 200,
-                      fit: BoxFit.fill),
-                )),
-            Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Padding(
-                  padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                  child: Text('Safety Level',
-                      style: TextStyle(height: 1, fontSize: 20)),
-                ),
-                Padding(
-                  padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                  child: RatingBar.builder(
-                    initialRating: 0,
-                    minRating: 1,
-                    direction: Axis.horizontal,
-                    allowHalfRating: true,
-                    itemCount: 5,
-                    itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                    itemSize: 25.0,
-                    itemBuilder: (context, _) => Icon(
-                      Icons.star,
-                      color: Colors.amber,
-                    ),
-                    onRatingUpdate: (rating) {
-                      safety = rating;
-                    },
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                  child: Text('Overall Experience',
-                      style: TextStyle(height: 1, fontSize: 20)),
-                ),
-                Padding(
-                    padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                    child: RatingBar.builder(
-                      initialRating: 0,
-                      minRating: 1,
-                      direction: Axis.horizontal,
-                      allowHalfRating: true,
-                      itemCount: 5,
-                      itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                      itemSize: 25.0,
-                      itemBuilder: (context, _) => Icon(
-                        Icons.star,
-                        color: Colors.lightBlue,
-                      ),
-                      onRatingUpdate: (rating) {
-                        overall = rating;
-                      },
-                    )),
-                Padding(
-                    padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                    child: TextField(
-                        controller: myController,
-                        decoration: InputDecoration(
-                          border: OutlineInputBorder(),
-                          labelText: 'Write your comment here',
-                          suffixIcon: IconButton(
-                            icon: Icon(Icons.help_center),
-                            onPressed: () {
-                              showDialog(
-                                context: context,
-                                builder: (BuildContext context) {
-                                  return AlertDialog(
-                                    title: Text('How should I make a review?'),
-                                    content: const Text('''
-You can determine the level of safety using these criteria:\n\
-1. The number of people in the place (crowdiness) \n
-2. Percentage of people (especially staffs) wearing masks \n 
-3. Whether hand sanitizers are well equipped or not \n
-4. Other policies enforced in that place (e.g. In cafeteria, people are not allowed to sit face-to-face.) 
-'''),
-                                    actions: [
-                                      FlatButton(
-                                        child: Text('Ok'),
-                                        onPressed: () {
-                                          Navigator.of(context).pop();
-                                          //Navigator.popUntil(context, ModalRoute.withName(Navigator.defaultRouteName));
-                                        },
-                                      ),
-                                    ],
-                                  );
-                                },
-                              );
-                            },
-                          ),
-                        ),
-                        maxLines: null)),
-                Padding(
-                    padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-                    child: RaisedButton(
-                        color: Colors.lightBlue,
-                        onPressed: () {
-                          content = myController.text;
-                          showDialog(
-                            context: context,
-                            builder: (BuildContext context) {
-                              return AlertDialog(
-                                title: Text('Submission Successful'),
-                                content: const Text('Thanks for your support!'),
-                                actions: [
-                                  FlatButton(
-                                    child: Text('Ok'),
-                                    onPressed: () {
-                                      //Navigator.of(context).pop();
-                                      Navigator.popUntil(
-                                          context,
-                                          ModalRoute.withName(
-                                              Navigator.defaultRouteName));
-                                    },
-                                  ),
-                                ],
-                              );
-                            },
-                          );
-                        },
-                        child: Icon(Icons.publish, color: Colors.white)))
-              ],
-            ),
-          ],
-        ));
   }
 }
